fix: branch organization for merger #30
