+++ src/views/tests/ComponentsView.vue
@@ -1,10 +1,11 @@
 <script setup>
 import ContainerDefault from '@/components/ContainerDefault.vue';
-import AditionalInfoComp from '@/components/AditionalInfoComp.vue';
+import AdditionalInfoComp from '@/components/AdditionalInfoComp.vue';
 </script>
 
 <template>
     <h1>Biblioteca de Componentes</h1>
+    
     <!-- <hr> -->
     <!-- <p>component name</p> -->
     <!-- <Component -->
@@ -16,9 +17,10 @@ import AditionalInfoComp from '@/components/AditionalInfoComp.vue';
         test
     </ContainerDefault>
     <hr>
-
+        
     <hr>
     <p>AditionalInfo</p>
-    <AditionalInfoComp :tittle="'Materiais de Apoio'" :description="'Que tal estudar com um quiz interativo ou talvez com um mapa menta. Clique no botão e aproveite nossos materias adicionais'" />
+    <AdditionalInfoComp :title="'Materiais de Apoio'" :description="'Que tal estudar com um quiz interativo ou talvez com um mapa menta. Clique no botão e aproveite nossos materias adicionais'" />
     <hr>
+    
 </template>
